Added image, sound, and level constants
The constants module was updated to include image, sound, and
level constants. This was done to avoid hard coding them elsewhere
in the package. It also means that they can be easily updated if
needed. Changing where game resources are located or changing
file names is made easier as only the constants module would
need to be updated. All modules affected by this change were
also updated accordingly.