Added loading of different color brick images
Code was added to load different colored brick images based on
the color indicated in the xml level file. This was done so that
levels can include both brick placement patterns AND brick color
patterns. This increases the visual appeal of the game.

This functionality was achieved by loading the brick colors from
the xml into a list of lists similar to the list of lists for
brick placement. This list was then added to the level dict. A
dict mapping a color to a filename was added in the constants
module in order know which image file to use for each color. The
brick sprite was updated to take the image filename as a param
to its init method in order to load the appropriate image.