Small change brick module
The brick module was modified so that the brick init method takes
a color string instead of a file name as a parameter. This makes
more sense than having to pass in a filename. It makes more sense
to create a brick with a certain color rather than with a certain
filename.