Updated background artwork
The background image was updated to a darker black/grey background
without the brick pattern. This was done to make the game easier
on the eyes. The brickwall background strained your eyesight
because of all the extra lines and patterns. The constants module
was also updated to reflect this change.