Fixed mouse button glitch in menu scenes
There was a small bug where the user could click the mouse when
not on a button then move the mouse over the button and the button
would think it got pressed. This was fixed by adding MOUSEBUTTONUP
event checks in the event loop and setting the pressed variable
accordingly. The list of draw rects was also updated to include
the level draw rect.