XML Parser was switched to lxml.
Parsing the xml level files is now done using the lxml python
library. This change was made because lxml provides the ability
to validate xml files against dtd's and schemas. This feature will
be used in the future to ensure that the level file being loaded
is formatted properly for the game.