Removed unnecessary cruft from level files.
XML id's and attributes were removed from the bricks/row/column
elements as they were not serving any real purpose. The code that
parses them is dynamic and does not rely on a set number of rows
or columns so missing rows/columns is not a big issue.

Other items that were removed were unecessary row/column elements
in levels that have less than ten rows. These items were
unnecessary due to how the levels are dynamically parsed and drawn
on the screen. It doesn't make sense to have an element whose only
purpose is to tell the game not to draw a brick there. It makes
more sense just to omit those elements.