Began level editor reimplementation
The basic framework for the level editor has been written using
the tkinter gui library. There is a file menu for opening and
saving level files and there is a quit button for exiting the
application. There is also three input boxes for the level name,
ball speed, and next level. The actual editor part for placing
bricks has yet to be written.