Updates and Fixes.
The issue of the brick images not displaying was fixed in the
editor/brick.py module by adding a reference to the image object
to the button object.

A 'New' option was added under the file menu for starting a new
blank level. It's functionality has not been implemented.

The text entry boxes are now filled with the default values of the
level that is loaded.