Updated level files.
The level files were updated to have a 10 rows of 10 columns
rather than just having rows where there were actually bricks.
A value of 'none' was given in the columns where there were no
bricks. This was done in order for the editor buttons to display
properly. If there isn't a complete set of rows in the xml files,
then only rows with bricks will show up in the editor.