Brick array updates on button press.
The brick array in brick.py is updated whenever a button is
pressed. Continually updating the array is easier than trying to
build an array when the level must be save. Since the array is
always up to date, the brick array only has to be passed to the
save level function.