switched to relative imports, removed utils
Imports are now all explicit relative imports. This was done so
that future updates to package names will be easier to implement.

The utils package was removed in favor of individual modules at the
breakout package level. This was done because the 2 modules in utils
did not need to be grouped under the same namespace. The constants
module was also renamed to config.py to better describe its use.